update stableJoeDayData on events, track deposit fees collected
