+++ subgraphs/stablejoe/stablejoe.rinkeby.yaml
@@ -31,4 +31,4 @@ dataSources:
           handler: handleDepositFeeChanged
         - event: Withdraw(indexed address,uint256)
           handler: handleWithdraw
-      file: ./src/mapping.ts
+      file: ./src/mappings/stablejoe.ts
